// features/individuals-portal/models/valuation-request-information-validation.ts
export class ValuationRequestInformationValidation {
  isFrist: boolean = false;
  isSecond: boolean = false;
  isThird: boolean = false;
  isMultiRealestate: boolean = false;
  VALUATION_TITLE_VALIDATION_MESSAGE: string = '';
  OWNERSHIP_NUMBER_VALIDATION_MESSAGE: string = '';
  PURPOSE_VALUATION_Id_VALIDATION_MESSAGE: string = '';
  REASON_VALUATION_DETAILS_VALIDATION_MESSAGE: string = '';
  AUTOMATIC_SELECTION_OFFER_DURATION: string = '';
  DYNAMIC_DISTRIBUTION_REQUEST_DURATION_MESSAGE: string = '';
  NUMBER_OF_REPORTS_MESSAGE: string = '';
  APPLICATION_TYPE_MESSAGE: string = '';
  CUSTOMER_ID_MESSAGE: string = '';
  DESCRIPTION_MESSAGE: string = '';
  REPORT_DELIVERY_DATE_MESSAGE: string = '';
  CANCELlATION_DURATION_VALUEATION_DATE: string = '';
  REAL_ESTATE_TYPE_MESSAGE: string = '';
  REAL_ESTATE_USAGE_MESSAGE: string = '';
  GOOGLE_LOCATION_X_MESSAGE: string = '';
  GOOGLE_LOCATION_Y_MESSAGE: string = '';
  REAL_ESTATE_AREA_MESSAGE: string = '';
  GOOGLE_FULL_ADDRESS_MESSAGE: string = '';
  GOOGLE_COUNTRY_NAME_MESSAGE: string = '';
  GOOGLE_CITY_NAME_MESSAGE: string = '';
  GOOGLE_DISTRICT_NAME_MESSAGE: string = '';
  GOOGLE_COUNTRY_EXTRA_MESSAGE: string = '';
  VALUATION_ORGANIZATION_MESSAGE: string = '';
}